requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
protobuf==3.12.2
pytube3==9.6.4
six==1.15.0
typing-extensions==3.7.4.2
Werkzeug==1.0.1
